Copy mboot.jar to parent dir
git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk@162885 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-mboot2/build.bat
@@ -1,13 +1,15 @@
 @echo off
- 
+
 set buildDir=target
-  
-rmdir /S/Q %buildDir%
-   
+
+if exist %buildDir% rmdir /S/Q %buildDir%
+
 mkdir %buildDir%
-    
+
 %JAVA_HOME%\bin\javac -d %buildDir% @sources.txt
-     
+
 cd %buildDir% 
 %JAVA_HOME%\bin\jar -cfm ..\mboot.jar ..\manifest.txt *
 cd ..
+
+cp mboot.jar ..
